First version of LogicStart Quickstart for the ZPUino
Switches control LED's
A mod file is played out the audio jack
An arcade graphic is displayed on the VGA output using the VGAZX video
adapter
Joystick controls counters on the 7-segment LED
Analog is measured and displayed on the 7-segment LED

Known Issues:

The Analog input is not working on properly on all channels.
This only works on the Papilio One 500K right now, we have not generated
a bit file to work on the Papilio One 250K yet.